refactor(connections): extract ConnectionFormBody with presetSourceId
Splits connection-form.tsx into a Dialog-free body (state, discovery WS,
stream table, save logic) and a thin Dialog wrapper, so the add-source
wizard's step 3 (D8) can reuse the connection UI without a nested dialog
and lock the source to the one just created.

- presetSourceId (new): create mode, shows a read-only source display and
  auto-runs discovery — used by the wizard.
- lockedSourceId (existing): create mode, keeps the picker visible but
  disabled and preselected — used by the Ingest source row's "Add
  connection", unchanged.
- DialogTitle/DialogDescription stay in the wrapper (not the body): they
  depend on Radix Dialog context and would throw if the body is ever
  rendered outside a <Dialog>, which is the whole point of this split.
